Change colid of sys.syscolumns to type smallint
Changed the data type of column colid of system view sys.syscolumns from int to smallint.

Task: BABEL-2146
Author: Sharu Goel <goelshar@amazon.com>

+++ contrib/babelfishpg_tsql/sql/sys_views.sql
@@ -1041,7 +1041,7 @@ SELECT out_name as name
   , out_max_length as length
   , 0::sys.tinyint as xprec
   , 0::sys.tinyint as xscale
-  , out_column_id as colid
+  , out_column_id::smallint as colid
   , 0::smallint as xoffset
   , 0::sys.tinyint as bitpos
   , 0::sys.tinyint as reserved
